wizard_tg_bot/requirements.txt
python-telegram-bot>=20.6
python-dotenv>=1.0.1
supabase>=2.3.0